Refactors NodeStdOutReporter into TerminalReporter, puts it in its own file.
